modernize(test): replace C-style casts with C++ cast operators
Convert all C-style casts in test files to the appropriate C++ cast
operator, as flagged by clang-tidy's google-readability-casting check.

- static_cast<T>(expr) for numeric narrowing conversions
  (e.g. (uint8_t)i → static_cast<uint8_t>(i))
- reinterpret_cast<T>(expr) for pointer reinterpret
  (e.g. (struct sockaddr*)&addr → reinterpret_cast<struct sockaddr *>(&addr))
- const_cast<T>(expr) for adding/removing const
  (e.g. (void*)const_char_ptr → const_cast<char *>(const_char_ptr))

All 406 warnings resolved. Verified with:
  clang-tidy -p build --checks='-*,google-readability-casting' libx/x/**/*_test.cpp

Tests pass on default build (8/8 ctest passed, 96s).
No behavior change — only cast syntax modernized.

Files changed: 43 (test files + shared test helpers)

+++ libx/x/buf/ring_test.cpp
@@ -79,7 +79,7 @@ TEST(xRingBuffer, Full) {
   size_t      cap = xRingBufferCap(rb);
 
   char data[16];
-  memset(data, 'X', (size_t)cap);
+  memset(data, 'X', cap);
   ASSERT_EQ(xRingBufferWrite(rb, data, cap), cap);
   EXPECT_TRUE(xRingBufferFull(rb));
